Cluster anaylsis, first draft
Compute cluster with optics version. So far, there is 6 clusters, one of them is discarded, being the outlier.
For each cluster, dataframe have been created holding the people in the cluster, and the title of movies
the participated in. One frame also contains the different ethnicties of the people whitin the clusters.
Plots have been created showing the most represented ethnicities within a cluster.